% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedges_g.R
\name{hedges_g}
\alias{hedges_g}
\title{Calculation of the effect size Hedge's g for group comparisons}
\usage{
hedges_g(data, n_experimental_group, n_control_group, mean_experimental_group,
  mean_control_group, sd_experimental_group, sd_control_group)
}
\arguments{
\item{data}{a list of values as returned by the significance test}

\item{n_experimental_group}{number of participants in the experimental group}

\item{n_control_group}{number of participants in the control group}

\item{mean_experimental_group}{mean of the values of the experimental group}

\item{mean_control_group}{mean of the values of the control group}

\item{sd_experimental_group}{standard deviation of the values of the experimental group}

\item{sd_control_group}{standard deviation of the values of the control group}

\item{N}{number of total participants}
}
\value{
Hedge's g for group comparisons
}
\description{
<ES estimate for group designs with a total sample larger than 20, comparing
two independent or dependent groups: Standardised mean difference calculated by
using the difference between the post-test means in the numerator of the equation
 and the standard deviation units in the denominator (pooled SDs are used to correct for possible
  upwards bias as especially problematic in samples smaller than 50). Standardisation
   permits direct comparisons across studies using the same index of effect.>
}
