% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_effects_d.R
\name{mixed_d}
\alias{mixed_d}
\title{Calculation of the effect size Hedge's d for categorical predictors in mixed effects models}
\usage{
mixed_d(cat_mixed_effects_model, t, n_observations_exp, n_observations_ctrl,
  n_experimental_group, n_control_group, k_parameters, sB, sE)
}
\arguments{
\item{cat_mixed_effects_model}{tested mixed effects model including only categorial predictors}

\item{t}{t-value of the predictor d is to be calculated for}

\item{n_observations_exp}{number of observations in the experimental group}

\item{n_observations_ctrl}{number of observations in the control group}

\item{n_experimental_group}{number of subjects in the experimental group}

\item{n_control_group}{number of subjects in the control group}

\item{k_parameters}{number of parameters (including the intercept)}

\item{sB}{between subject's or between group variance}

\item{sE}{within individual or group variance}
}
\value{
hedge's d for categorical predictors in mixed effects models
}
\description{
<ES for mixed-effects models including only predictors which are categorical:
 For mixed-effects models as used if data is hierarchical and/or includes repeated-measures,
  t-values can be used to approximate d statistics.>
}
