% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_d.R
\name{multiple_d}
\alias{multiple_d}
\title{Calculation of the effect size Hedges's d for continuous and categorical predictors in GLM and multiple linear regression models}
\usage{
multiple_d(model, t, df, n_experimental, n_control)
}
\arguments{
\item{model}{a list of values as returned testing a general linear or multiple-regression model}

\item{t}{t-value of the respective predictor}

\item{df}{degrees of freedom in the model}

\item{n_experimental}{number of participants in the experimental group}

\item{n_control}{number of participants in the control group}
}
\value{
Hedge's d for continuous and categorical predictors in GLM and multiple
 linear regression models
}
\description{
<ES for continuous and categorical predictors in general linear and
multiple regression models, calculated using t-values (obtained from a difference between estimates,
e.g. means or slopes, divided by the standard error of the differences - almost all statistical software
provides t values when a statistical model is constructed). The t-values obtained for groups or
categories in a predictor variable can be used for calculating d.>
}
