% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_r.R
\name{multiple_r}
\alias{multiple_r}
\title{Calculation of the effect size r for continuous predictors in GLM and multiple regression models}
\usage{
multiple_r(model, t, df)
}
\arguments{
\item{model}{tested general linear or multiple-regression models with continuous predictors}

\item{t}{t-value of the continuous predictor d is to be calculated for}

\item{df}{degrees of freedom in the model}
}
\value{
r statistics for continuous predictors in GLM and multiple regression models
}
\description{
<ES for continuous predictors in general linear and multiple
regression models, calculated using t values, which are generally obtained from
 a difference between estimates (e.g. means or slopes) divided by the standard
 error of the differences - almost all statistical software provides t values
 when a statistical model is constructed. The t values obtained for a continuous
  predictor variable can be used for calculating r.>
}
