% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cohens_d.R
\name{run_cohens_d}
\alias{run_cohens_d}
\title{Wizard assisted calculation of the effect size Cohen's d for group comparisons with small samples}
\usage{
run_cohens_d()
}
\arguments{
\item{N}{number of total subjects tested}

\item{mean_experimental}{mean of the values observed in the tested experimental group}

\item{mean_control}{mean of the values observed in the tested control group}

\item{sd_experimental}{standard deviation of the values observed in the tested experimental group}

\item{sd_control}{standard deviation of the values observed in the tested control group}
}
\value{
Cohens' d after group comparisons
}
\description{
<Effect size for group designs including very small samples (N < 20), comparing two independent
 or dependent groups: Standardized mean difference calculated using the difference between post-test means in the
  numerator of the equation and the standard deviation units in the denominator (pooled SDs are used to correct
  for possible upwards bias as problematic in very small samples). Standardisation permits direct comparisons across
  studies using the same index of effect.>
}
