% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multiple_r.R
\name{run_multiple_r}
\alias{run_multiple_r}
\title{Wizard assisted calculation of the effect size r for continuous predictors in GLM and multiple regression models}
\usage{
run_multiple_r()
}
\arguments{
\item{t}{t-value of the continuous predictor d is to be calculated for}

\item{df}{degrees of freedom in the model}
}
\value{
r statistics for continuous predictors in GLM and multiple regression models
}
\description{
<ES for continuous predictors in general linear and multiple
regression models, calculated using t values, which are generally obtained from
 a difference between estimates (e.g. means or slopes) divided by the standard
 error of the differences - almost all statistical software provides t values
 when a statistical model is constructed. The t values obtained for a continuous
  predictor variable can be used for calculating r.>
}
