% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_partial_correlation_r.R
\name{run_partial_r}
\alias{run_partial_r}
\title{Wizard assisted calculation of a partial correlation coefficient}
\usage{
run_partial_r()
}
\arguments{
\item{r_1_2}{correlation between variable 1 and variable 2}

\item{r_2_3}{correlation between variable 2 and variable 3}

\item{r_1_3}{correlation between variable 1 and variable 3}
}
\value{
Partial correlation coefficient
}
\description{
<Partial correlation coefficient: The partial correlation between y
and x1, controlling for x2, as numerical equivalent to the correlation between
the residuals of the regression of y on x2 and the residuals of the regression of
 x1 on x2. The partial coefficient for a given predictor removes the variance
 explained by other predictor variables from both variables, and then quantifies
  the remaining correlation.>
}
