% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_wizard.R
\name{run_wizard}
\alias{run_wizard}
\title{Wizard for the selection of an adequate effect size}
\usage{
run_wizard()
}
\description{
<This calculation wizard guides the user through the selection of the adequate effect size estimate.
The adequate efect size is selected based on several questions asked via dialog boxes. Subsequently, the
selected effect size function included in the package is prompted and/or the adequate function recommended.
The functions for calculating the effect sizes contain themselves a wizard that inquires the necessary data to
calculate the effect size.
}
