% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_r.R
\name{wilcox_r}
\alias{wilcox_r}
\title{Calculation of the effect size r after group comparisons with Wilcoxon rank-sum/signed-rank test}
\usage{
wilcox_r(wilcoxModel, N)
}
\arguments{
\item{N}{number of total subjects tested}
}
\value{
Effect size estimate r calculated by converting z-values
}
\description{
<Effect size for group designs comparing two independent or dependent groups concerning their
medians, calculated by converting z-values into an effect size estimate r.
#' @param wilcoxModel a list of values as returned by Wilcoxon test
}
